Fix all non-suppressed scss-lint issues
Mostly this involved dropping outdated vendor prefixes, sorting
properties and then I dropped the whole popup that was used for
the navigation menu mobile devices (the menu itself looks ugly
anyway and the SCSS for it was nested 4 levels deep and had various
other lint issues. I plan to add a new nicer menu for mobile later on).